[myanmar] Add dump-myanmar-data
